headless_lms_base

Module prelude_base_and_re_exports

Re-exports

pub use crate::config::ApplicationConfiguration;

pub use crate::config::AzureBlobStorageConfiguration;

pub use crate::config::AzureChatbotConfiguration;

pub use crate::config::AzureConfiguration;

pub use crate::config::AzureSearchConfiguration;

pub use crate::config::OAuthServerConfiguration;

pub use crate::error::backend_error::BackendError;

pub use anyhow;

pub use chrono;

pub use futures;

pub use serde_json;

pub use tokio;

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

info

Constructs an event at the info level.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

DateTime

ISO 8601 combined date and time with time zone.

Url

A parsed URL record.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Uuid

A Universally Unique Identifier (UUID).

Traits

Context

Provides the context method for Result.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Derive Macros

Deserialize

Serialize